Cleanup of the float -> double switch. (#820)
Previously to avoid having the code base differing too much from
the original ITS reference implementation, we defined a global
  #define float double
in the ehata.h header file.
Obviously this is very hack and can fail depending on the order of
inclusion of ehata.h, or if t is forgotten in one of the cpp file.

The proper way is now implemented by directly changing all `float`
occurence with `double`.
This is in line with the change made by ITS on their reference
implementation at:
https://github.com/NTIA/ehata/commit/df2b5802f1444b77ffadf4a4301c5b2fd7be691c

+++ src/harness/reference_models/propagation/ehata/its/FineRollingHillyTerrainCorectionFactor.cpp
@@ -10,16 +10,16 @@
 *       interValues->deltah__meter : terrain irregularity parameter
 *       h_m_gnd__meter : mobile ground height, in meters
 *   Return:
-*       [float] : correction factor
+*       [double] : correction factor
 */
-float FineRollingHillyTerrainCorectionFactor(InterValues *interValues, float h_m_gnd__meter)
+double FineRollingHillyTerrainCorectionFactor(InterValues *interValues, double h_m_gnd__meter)
 {
-    float a = -11.728795;
-    float b = 15.544272;
-    float c = -1.8154766;
+    double a = -11.728795;
+    double b = 15.544272;
+    double c = -1.8154766;
 
-    float deltah_use;
-    float K_h;
+    double deltah_use;
+    double K_h;
 
     // deltaH must be at least 10 meters
     if (interValues->deltah__meter < 10.0)
